+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Send 204 Status instead of empty response when querying data, [PR-1125](https://github.com/reductstore/reductstore/pull/1125)
 - Prevent race condition between sync and remove operations, [PR-1127](https://github.com/reductstore/reductstore/pull/1127)
 - Allow empty batch v2 in parsing, [PR-1134](https://github.com/reductstore/reductstore/pull/1134)
+- Remove entries and buckets in the background and per block, [PR-1135](https://github.com/reductstore/reductstore/pull/1135)
 
 ## 1.17.9 - 2025-01-16